add and fix integration tests
