+++ src/integration-test/java/uk/gov/hmcts/juror/api/moj/controller/JurorRecordControllerITest.java
@@ -131,7 +131,6 @@ import static org.springframework.http.HttpMethod.POST;
 import static uk.gov.hmcts.juror.api.juror.domain.ProcessingStatus.AWAITING_COURT_REPLY;
 import static uk.gov.hmcts.juror.api.moj.exception.MojException.BusinessRuleViolation.ErrorCode.FAILED_TO_ATTEND_HAS_ATTENDANCE_RECORD;
 import static uk.gov.hmcts.juror.api.moj.exception.MojException.BusinessRuleViolation.ErrorCode.FAILED_TO_ATTEND_HAS_COMPLETION_DATE;
-import static uk.gov.hmcts.juror.api.moj.exception.MojException.BusinessRuleViolation.ErrorCode.JUROR_STATUS_MUST_BE_FAILED_TO_ATTEND;
 import static uk.gov.hmcts.juror.api.moj.exception.MojException.BusinessRuleViolation.ErrorCode.JUROR_STATUS_MUST_BE_RESPONDED;
 
 
@@ -4145,151 +4144,6 @@ class JurorRecordControllerITest extends AbstractIntegrationTest {
             JurorPool jurorPool =
                 jurorPoolRepository.findByJurorJurorNumberAndPoolPoolNumber(JUROR_NUMBER, POOL_NUMBER);
             assertEquals(IJurorStatus.RESPONDED, jurorPool.getStatus().getStatus(),
-                "Juror pool status should not change");
-
-            List<JurorHistory> jurorHistories = jurorHistoryRepository.findByJurorNumberOrderById(JUROR_NUMBER);
-            assertEquals(0, jurorHistories.size(),
-                "No new history entry as request should be rejected before processing");
-        }
-    }
-
-    @Nested
-    @DisplayName("PATCH " + UndoUpdateJurorToFailedToAttend.URL)
-    @Sql({"/db/mod/truncate.sql",
-        "/db/JurorRecordControllerITest_failedToAttend_undo_typical.sql"})
-    class UndoUpdateJurorToFailedToAttend {
-
-        private static final String URL = BASE_URL + "/failed-to-attend/undo";
-
-        private static final String JUROR_NUMBER = "641500005";
-        private static final String POOL_NUMBER = "415220901";
-
-        private JurorNumberAndPoolNumberDto createDto(String jurorNumber, String poolNumber) {
-            JurorNumberAndPoolNumberDto dto = new JurorNumberAndPoolNumberDto();
-            dto.setJurorNumber(jurorNumber);
-            dto.setPoolNumber(poolNumber);
-            return dto;
-        }
-
-        @Test
-        void positiveTypical() {
-            setAuthorization("COURT_USER", "415", UserType.COURT, Role.SENIOR_JUROR_OFFICER);
-            JurorNumberAndPoolNumberDto dto = createDto(JUROR_NUMBER, POOL_NUMBER);
-            ResponseEntity<Void> response =
-                restTemplate.exchange(
-                    new RequestEntity<>(dto, httpHeaders, HttpMethod.PATCH, URI.create(URL)), Void.class);
-
-            assertThat(response).isNotNull();
-            assertThat(response.getStatusCode()).isEqualTo(HttpStatus.ACCEPTED);
-
-            JurorPool jurorPool =
-                jurorPoolRepository.findByJurorJurorNumberAndPoolPoolNumber(JUROR_NUMBER, POOL_NUMBER);
-            assertEquals(IJurorStatus.RESPONDED, jurorPool.getStatus().getStatus(),
-                "Juror pool status should be responded");
-
-            List<JurorHistory> jurorHistories = jurorHistoryRepository.findByJurorNumberOrderById(JUROR_NUMBER);
-            assertEquals(1, jurorHistories.size(), "Should only be one history entry");
-            JurorHistory jurorHistory = jurorHistories.get(0);
-            assertEquals(POOL_NUMBER, jurorHistory.getPoolNumber(), "Pool number should match");
-            assertEquals(JUROR_NUMBER, jurorHistory.getJurorNumber(), "Juror number should match");
-            assertEquals("COURT_USER", jurorHistory.getCreatedBy(), "User id should match");
-            assertEquals(HistoryCodeMod.FAILED_TO_ATTEND, jurorHistory.getHistoryCode(), "History code should match");
-            assertEquals("FTA status removed", jurorHistory.getOtherInformation(),
-                "Info should match");
-        }
-
-        @Test
-        void negativeNotFound() {
-            setAuthorization("COURT_USER", "415", UserType.COURT, Role.SENIOR_JUROR_OFFICER);
-            JurorNumberAndPoolNumberDto dto = createDto("123456789", POOL_NUMBER);
-            ResponseEntity<String> response =
-                restTemplate.exchange(
-                    new RequestEntity<>(dto, httpHeaders, HttpMethod.PATCH, URI.create(URL)), String.class);
-            assertErrorResponse(response,
-                HttpStatus.NOT_FOUND,
-                URL,
-                MojException.NotFound.class,
-                "Juror number 123456789 not found in pool " + POOL_NUMBER);
-        }
-
-        @Test
-        void negativeNotFailedToRespond() {
-            final String jurorNumber = "641500004";
-            setAuthorization("COURT_USER", "415", UserType.COURT, Role.SENIOR_JUROR_OFFICER);
-            JurorNumberAndPoolNumberDto dto = createDto(jurorNumber, POOL_NUMBER);
-
-            ResponseEntity<String> response =
-                restTemplate.exchange(
-                    new RequestEntity<>(dto, httpHeaders, HttpMethod.PATCH, URI.create(URL)), String.class);
-            assertBusinessRuleViolation(response,
-                "Juror status must be failed to attend in order to undo the failed to attend status.",
-                JUROR_STATUS_MUST_BE_FAILED_TO_ATTEND
-            );
-
-            JurorPool jurorPool = jurorPoolRepository.findByJurorJurorNumberAndPoolPoolNumber(jurorNumber, POOL_NUMBER);
-            assertEquals(IJurorStatus.RESPONDED, jurorPool.getStatus().getStatus(),
-                "Juror pool status should not change");
-
-            List<JurorHistory> jurorHistories = jurorHistoryRepository.findByJurorNumberOrderById(jurorNumber);
-            assertEquals(0, jurorHistories.size(),
-                "No new history entry as request should be rejected before processing");
-        }
-
-
-        @Test
-        void negativeInvalidPayload() {
-            setAuthorization("COURT_USER", "415", UserType.COURT, Role.SENIOR_JUROR_OFFICER);
-            JurorNumberAndPoolNumberDto dto = createDto("INVALID", POOL_NUMBER);
-            ResponseEntity<String> response =
-                restTemplate.exchange(
-                    new RequestEntity<>(dto, httpHeaders, HttpMethod.PATCH, URI.create(URL)), String.class);
-            assertInvalidPayload(response,
-                new RestResponseEntityExceptionHandler.FieldError("jurorNumber", "must match \"^\\d{9}$\""));
-
-
-            JurorPool jurorPool =
-                jurorPoolRepository.findByJurorJurorNumberAndPoolPoolNumber(JUROR_NUMBER, POOL_NUMBER);
-            assertEquals(IJurorStatus.FAILED_TO_ATTEND, jurorPool.getStatus().getStatus(),
-                "Juror pool status should not change");
-
-            List<JurorHistory> jurorHistories = jurorHistoryRepository.findByJurorNumberOrderById(JUROR_NUMBER);
-            assertEquals(0, jurorHistories.size(),
-                "No new history entry as request should be rejected before processing");
-        }
-
-        @Test
-        void negativeUnauthorisedWrongLevel() {
-            setAuthorization("COURT_USER", "415", UserType.COURT);
-            JurorNumberAndPoolNumberDto dto = createDto(JUROR_NUMBER, POOL_NUMBER);
-            ResponseEntity<String> response =
-                restTemplate.exchange(
-                    new RequestEntity<>(dto, httpHeaders, HttpMethod.PATCH, URI.create(URL)), String.class);
-
-            assertForbiddenResponse(response, URL);
-
-            JurorPool jurorPool =
-                jurorPoolRepository.findByJurorJurorNumberAndPoolPoolNumber(JUROR_NUMBER, POOL_NUMBER);
-            assertEquals(IJurorStatus.FAILED_TO_ATTEND, jurorPool.getStatus().getStatus(),
-                "Juror pool status should not change");
-
-            List<JurorHistory> jurorHistories = jurorHistoryRepository.findByJurorNumberOrderById(JUROR_NUMBER);
-            assertEquals(0, jurorHistories.size(),
-                "No new history entry as request should be rejected before processing");
-        }
-
-        @Test
-        void negativeUnauthorisedBureau() {
-            setAuthorization("BUREAU_USER", "400", UserType.BUREAU);
-            JurorNumberAndPoolNumberDto dto = createDto(JUROR_NUMBER, POOL_NUMBER);
-            ResponseEntity<String> response =
-                restTemplate.exchange(
-                    new RequestEntity<>(dto, httpHeaders, HttpMethod.PATCH, URI.create(URL)), String.class);
-
-            assertForbiddenResponse(response, URL);
-
-            JurorPool jurorPool =
-                jurorPoolRepository.findByJurorJurorNumberAndPoolPoolNumber(JUROR_NUMBER, POOL_NUMBER);
-            assertEquals(IJurorStatus.FAILED_TO_ATTEND, jurorPool.getStatus().getStatus(),
                 "Juror pool status should not change");
 
             List<JurorHistory> jurorHistories = jurorHistoryRepository.findByJurorNumberOrderById(JUROR_NUMBER);
